refactor: extract AnalyzeRepoDeep, add ReportWriterPort/MarkdownReportWriter
Sixth extraction in the hexagon-cleanup plan, and the last of the five
planned use cases. Pulls the deep-analysis flow out of agent.ts:

- buildRepoContext, detectStackSignals, buildStructureOverview,
  detectSetupSignals, buildRiskDetails, loadScoutSelectionContext,
  writeAnalysisReport -> AnalyzeRepoDeep.execute(repo) -> RepoContext

Required a real port, not just a function move: agent.ts was writing
reports/REPO_ANALYSIS.md and reading/writing reports/REPO_SCOUT_RESULTS.md
via raw mkdir/readFile/writeFile. That's exactly the I/O the hexagon
docs say should sit behind a port. Added:

- ReportWriterPort (writeAnalysisReport, writeScoutReport, readScoutReport)
- MarkdownReportWriter implements it

The scout-report table parsing (loadScoutSelectionContext) split into
parseScoutSelectionContext — pure text parsing, now testable without
touching a filesystem — plus the port's readScoutReport() for the
actual file read. buildRepoContext was typed against the concrete
GithubAdapter class rather than the RepoApiPort interface it actually
needs; fixed as part of the move, since domain code can't depend on a
concrete adapter.

writeScoutReport (agent.ts, builds the shortlist markdown table) now
takes the same ReportWriterPort instead of calling writeFile directly
— it wasn't moving into AnalyzeRepoDeep this round (different flow,
still tied to the CLI-local RankedShortlistItem type), but leaving its
write on raw fs while the read it feeds went through the port would
have been an inconsistent halfway state.

agent.ts no longer imports node:fs/promises at all — every remaining
I/O in the file (session, both reports) now goes through a port.
1331 -> 1021 lines.

7 new tests (3 for parseScoutSelectionContext — match found, no match,
non-numeric score; 4 for AnalyzeRepoDeep.execute against mocked
RepoApiPort/AnalyzeRepo/ReportWriterPort — full context assembly,
scout-context correlation present and absent, risk-flag correctness).
Full suite: 72/72 green, tsc --noEmit clean.

Remaining on the hexagon-cleanup plan: LlmPort/OpenAiAdapter (collapses
QueryTranslator + AiBrain's duplicated OpenAI plumbing) — the only
piece not done in this pass.
